Fix handling --verify option validation
While client code replaces filename with /dev/fd/FD_NUM, server code
should check this and disallow any other value. The check is there, but
it's buggy - only prints a message but do not interrupt processing.

The result is that malicious client could point `gpg2 --verify` to
arbitrary file in the backend domain. But fortunatelly this doesn't
allow leaking any secret from the backend domain, because gpg2 verify if
the file is really a signature and prints an error if it isn't.

Fixes QubesOS/qubes-issues#2165