Don't mix stdout and --output content
When `--status-fd 1` and `--output` are given at the same time, it
should output status on stdout and write output data to a file. But
since --output is implemented as stdout redirection in a wrapper script,
this isn't the case.

Fix this by redirecting `--status-fd 1` to some high file descriptor,
which later gets redirected back to stdout - after actual data is being
redirected to a file.

Fixes QubesOS/qubes-issues#2057