Emulate --log-file with --logger-fd
According to documentation, those two are equivalent with exception of
using existing file descriptor instead of named file. This allows to
support --log-file option with a file on client-side, instead of
allowing write to arbitrary file on the backend side.
The idea is simple: in qubes-gpg-client-wrapper open a new file
descriptor (62) on the client side, redirect it to a file pointed by
--log-file, and instruct actual qubes-gpg-client to use --logger-fd 62
instead.
The commit also contains a test for this.

Fixes QubesOS/qubes-issues#3989