Wait for GUI session before asking the user for confirmation
qubes.Gpg service may be called before X server is started in the VM.
This will result in a failure and be treated as user rejecting the
access.

One option to fix this would be waiting for GUI session before (all)
qubes.Gpg calls. But:
1) this feature is built into qrexec in Qubes R4.0 only, for R3.2 it
would need to be manual qubes.WaitForSession call anyway
2) this would break the cases where confirmation is provided by some
other means (like in qubes-infrastructure)

Since manual call to qubes.WaitForSession is needed anyway, add it only
when really going to ask for confirmation.

Fixes QubesOS/qubes-issues#4528