Accept --personal-{cipher,...}-preferences with option-argument
The options --personal-cipher-preferences,
--personal-compress-preferences and --personal-digest-preferences were
incorrectly treated as having no option-arguments leading to either an
error message (when using long argument syntax) or the option-argument
being treated as an operand (which e.g. in the case of --clearsign would
be a filename).  This commit updates the option struct in gpg-common.h
to treat those options as having one option-argument.