tests: master key sign-only + subkey
Version 78.10.2 started supporting offline master keys. This change
broke the configuration where the master key is offline and has no
subkeys. This is a non-default PGP key generation and hence has
little change users are in this situation.

The commit generate a master key with signing-capabilities only
and a subkey for encryption

Fixes QubesOS/qubes-issues/issues/6712