tests: manage thunderbird through dedicated class
previously thunderbird was started, closed ad-hoc in various
places. This made it cumbersome to restart the app at any moment
(in the future if the present setup function fails it would have
to be restarted). The main reason was that to the process and to
the dogtail references of Thunderbird would be lost.

This way we have a dedicated class to keep track of state changes
to the application.