Reject abbreviated long options
getopt_long(3) allows long options to be abbreviated if they are
unambiguous or if they match the name of an option exactly.  This was
one of several sources of QSB#71.  Furthermore, GPG may add new long
options in the future, and some of these may be substrings of existing
long options.  This would create new QSB#71-style vulnerabilities.
Finally, abbreviated options are not handled properly by
qubes-gpg-client-wrapper.

To prevent abbreviated options from being a problem in the future, this
commit adds an explicit check to reject them.  I tested this with
gpg-client, as that program is easier to test, but as the parsing code
is identical in both cases, this ensures that gpg-server will also
reject abbreviated options.