Fix off-by-one in fd path replacement
If I had a GPG signature file at /tmp/sig that I wanted to verify,
qubes-gpg-client --verify /tmp/sig would fail with a very confusing
error message:

	$ qubes-gpg-client --verify /tmp/sig
	--verify with filename allowed only on the client side
	EOF

This happened because of an off-by-one error — untrusted_sig_path_size
includes the trailing null byte, while the return value of snprintf
does not.  Therefore, qubes-gpg-client would think that the full file
descriptor path had been written to untrusted_sig_path, even though
the final byte would not have been, so untrusted_sig_path would have
the value "/dev/fd/".  Alternatively, if a file descriptor greater
than ten was being used, and the path to the signature file was one
byte longer, the last decimal digit of the file descriptor number
would be omitted — /dev/fd/1 instead of /dev/fd/10, for example.

So we need to add one to the return value of snprintf to take the
trailing null byte into account before comparing it with
untrusted_sig_path_size.  After making this change:

	$ qubes-gpg-client --verify /tmp/sig
	Failed to fit /dev/fd/3 in place of signature path

Which is the expected behaviour.

This was not a security issue, because the bug was client-side, and
Split GPG design considers the client untrusted anyway.