Make fd path replacement length check clearer
The previous length checking code was confusing.  It used two variables,
untrusted_sig_path_size, and written.  untrusted_sig_path_size included
a trailing null byte, while written did not include the trailing null
byte.  This inconsistency was responsible for the bug I fixed in the
previous commit.  Additionally, "written" is a very confusing name,
because the return value of snprintf is not the number of bytes written,
it's the length of the string that it tried to write -- if the buffer
was five bytes long, and the file descriptor path was nine characters
long (excluding null terminator), then snprintf would return 9, even
though it only wrote five bytes.

Here, I've refactored the code slightly.  I renamed "written" to
"fd_path_len", which better describes what it is, and I've renamed
untrusted_sig_path_size to untrusted_sig_path_len, and made it exclude
the trailing null byte to match fd_path_len.  That way, there's no
longer a requirement to add one to fd_path_len before comparing it with
untrusted_sig_path_len.  I chose to exclude the trailing null byte from
both rather than include it for both because excluding it preserves
the property that fd_path_len will be negative if snprintf returned an
error, which I think is important for clarity.