tests: fix thunderbird IMAP password not showing
In tests like https://openqa.qubes-os.org/tests/26462#step/TC_10_Thunderbird_whonix-ws-16/2
whonix-based  appvms (thunderbird 91) they would try to click on
"Get Messages" in order for the IMAP password prompt to show but
this would not happen.

When debugging the password prompt would only show almost a minute
after. The reason for this was due to whonix's time jumps from
the time randomization services.

The fix was to avoid restarting dovecot and instead just reloading
its configuration -- this way it doesn't realize the time jump.