Properly validate file descriptor arguments
The previous code did not properly validate file descriptor arguments.
I was not able to figure out a way to exploit the lack of proper
validation, but it is definitely a bug.

Add a new function for parsing strings as file descriptors and use it in
both places where this is necessary.  This was tested manually, and will
be tested automatically as part of future split-gpg unit tests.