Avoid closing a needed file descriptor
The file descriptor remapping code had a fatal flaw: if the read side of
one of the pipes that would be used for GPG's output had the same number
as the argument to --command-fd, it would wind up being closed along
with the pipe read ends.  Fix this by ensuring that every file
descriptor that GPG will use is already open before creating the pipes.
ensuring that the kernel will not use any of GPG's file descriptors for
the pipes.  Since the set of file descriptors used by the pipes and by
GPG are now totally disjoint, the remap table is pointless and can be
eliminated.  Furthermore, closing FDs 0, 1, and 2 can be avoided, as
they will be dup2()'d over anyway.