tests: update Evolution test for newer settings dialog layout
The settings dialog has settings category buttons (with categories name)
not connected to actual settings tabs at the a11y level. Worse - they
are not even in the same order. We use heuristic to find the right tab,
by looking for a widget that is only on the "Accounts" tab.

There is no more 'Open Online Accounts' button, look for 'Account Name'
header instead.

QubesOS/qubes-issues#7475