Avoid hang due to premature file descriptor close
The pipe remapping code relies on all of the pipes being created at file
descriptors that are not their final locations.  This requires that all
of the file descriptors that *are* their final locations must be open
when pipe2(2) is called.  Closing the file descriptor to /dev/null
before creating the pipes could violate this invariant.  Fix the bug by
not closing any file descriptors until after all of the pipes have been
created.

Fixes QubesOS/qubes-issues#7530.