Use common code for file descriptor lists
Previously, both add_arg_to_fd_list() and handle_opt_verify() had their
own logic for adding file descriptors to lists.  Replace both with a
common add_fd_to_list() function.  As a bonus, the requirement that
/dev/fd/%d fit in place of the signature path is now lifted, by
replacing the pointer to the argument instead of overwriting the
argument in-place.  Finally, require the file descriptor for --verify on
the server side to be at least 3, to ensure that it is distinct from all
standard streams.