Refuse to use the same FD for both reading and writing
split-gpg1 uses pipes for all I/O to and from GnuPG.  Pipes are
unidirectional, so it is not possible for the same FD to be used for
both reading and writing.  The likely result is a hang or a confusing
"Bad file descriptor" error.  This is especially true because the
same call could actually be perfectly valid; consider the case where one
of the file descriptors is a socket.

To prevent this, explicitly check for using a file descriptor in both
directions and emit an error message.  If the caller is trying to write
to stdin, or to read from stdout or stderr, mention that specifically.