Disallow --command-fd
--command-fd appears safe, and indeed for a while it was thought to be
safe.  However, it turns out that GnuPG trusts the data it receives over
this file descriptor.  For instance, when not in batch mode, it will
prompt for a signed file if none is provided.  Also, it is trivial to
make GnuPG use a lot of CPU and memory, and it will eventually have a
(probably non-exploitable) integer overflow bug on really really long
input lines.

To avoid this problem, disallow --command-fd entirely.  It is used by
reprepro, so ignore it in the wrapper script.  Most of the operations
that require --command-fd, such as key generation and editing, are not
supported by split-gpg1 anyway.