Prevent GnuPG from using a closed file descriptor
If --status-fd, --logger-fd, or --attribute-fd is given more than once,
GnuPG will (with a few exceptions) close the file descriptor given the
first time.  If this is a standard stream, GnuPG will then continue to
use the stream.  This allows an attacker to corrupt the trust database
or send arbitrary, unsanitized commands to the agent.

Thankfully, passing any of these arguments more than once is likely a
bug in the calling program, so we do not need to support it.  Just
return an error in this case.