Properly NULL-terminate argument list
The code failed to properly NULL-terminate argument lists, causing the
server to treat uninitialized stack memory as a pointer.  The bug can be
reproduced using Clang's -ftrivial-auto-var-init=pattern; a server
compiled with this flag will crash on any input.

This is not a security problem.  The uninitialized pointer always turns
out to be NULL in practice, which is why the problem has been undetected
for so long.  Furthermore, if the pointer turns out to *not* be NULL, a
subsequent check will detect this and call abort(), safely terminating
the process.  Theoretically, a compiler could exploit the undefined
behavior to do something bad, but there is no evidence of this happening
in practice.

Fix this problem by replacing the on-stack arrays with static arrays,
which are always zero-initialized in C.  This also saves over 16K of
stack space.