Allow exporting public keyring backups
This will ease migration to split-gpg2, as users can now export a full
backup of their public keyring using split-gpg1.  split-gpg2 uses a
client-side keyring, so it is necessary to move the keyring from the
server to the client, and this is the simplest method.

--export-options=export-dane is not allowed as it brings in a bit of
extra code in GnuPG and can be done once the keyring has been imported
on the client.

Fixes QubesOS/qubes-issues#7588.