Avoid UBSAN splat on max-length input
If the input length was exactly 1024 bytes, the NUL terminator would be
written one past the end of untrusted_hdr->command.  This does not cause
problems in practice because that byte is in allocated memory, but it is
undefined behavior and UBSAN catches it.  Instead of writing a NUL
terminator, require that the client NUL-terminate the message itself,
and consider the NUL terminator to be part of the message sent by the
client.  That is, the length of the message is 1 more than the len field
in the header struct, which is the same computation used on the client
side.